It seems that the delay is not as big as I make it to be.
When I ran the old code from the server, the time did not change much.
The new code is slightly faster.  I think the bottle neck is still my parser
and I do not know how to make it faster.  Perhaps time can tell.

Why is it slowing down?  I believe it has to do with object creation.
I might be created to many objects.  String objects and array splitting
might be the culprit.